+++ src/components/Devices/PairDeviceBottomSheet.tsx
@@ -32,13 +32,13 @@ import useNoticeStore from '@/stores/notice';
 import useSettingsStore from '@/stores/settings';
 import useToastStore from '@/stores/toast';
 
-const detectLogger = log.extend(`[detect]`);
+const pairLogger = log.extend(`[pair-device]`);
 
-const FETCH_DEVICES_MAX_ATTEMPTS_COUNT = 10;
+const VERIFY_ATTENDING_DEVICE_MAX_ATTEMPTS_COUNT = 10;
 const FETCH_DEVICE_INFO_MAX_ATTEMPTS_COUNT = 10;
 const RETRY_BASE_DELAY_IN_MS = 1000;
 
-const DetectDeviceBottomSheet = ({
+const PairDeviceBottomSheet = ({
   style,
   onClose,
 }: {
@@ -58,9 +58,15 @@ const DetectDeviceBottomSheet = ({
 
   const [isReachingService, setReachingService] = useState<boolean>(false);
   const [isServiceReachable, setServiceReachable] = useState<boolean>(false);
+  const [fetchDeviveInfoAttemptsCount, setFetchDeviveInfoAttemptsCount] = useState<number>(0);
+  const [fetchDeviceTimeoutHandle, setFetchDeviceTimeoutHandle] = useState<NodeJS.Timeout | null>(
+    null,
+  );
   const [macAddress, setMacAddress] = useState<string | null>(null);
   const [name, setName] = useState<string | null>(null);
   const [addedDevice, setAddedDevice] = useState<ApiMemberDevice | null>(null);
+  const [verifyingAttemptsCount, setVerifyingAttemptsCount] = useState<number>(0);
+  const [verifyingTimeoutHandle, setVerifyingTimeoutHandle] = useState<NodeJS.Timeout | null>(null);
   const [verifiedDevice, setVerifiedDevice] = useState<ApiMemberDevice | null>(null);
 
   const [isAnimating, setAnimating] = useState<boolean>(false);
@@ -70,27 +76,36 @@ const DetectDeviceBottomSheet = ({
 
   // https://stackoverflow.com/a/65805464
   const verifyAttendingDevice = useCallback(
-    async (memberId: string, deviceId: string, count = 0) => {
+    async (memberId: string, deviceId: string, count = 1) => {
       return new Promise((resolve, reject) => {
-        if (count > FETCH_DEVICES_MAX_ATTEMPTS_COUNT) {
-          throw new Error(t('devices.add.onVerifyDevice.tooManyAttempts'));
-        }
+        const handle = setTimeout(() => {
+          pairLogger.debug(`verifyAttendingDevice ${deviceId} attempt ${count}`);
+          if (count > VERIFY_ATTENDING_DEVICE_MAX_ATTEMPTS_COUNT) {
+            return reject(
+              Error(
+                t('devices.add.onVerifyDevice.tooManyAttempts', {
+                  count: VERIFY_ATTENDING_DEVICE_MAX_ATTEMPTS_COUNT,
+                }),
+              ),
+            );
+          }
 
-        if (abortController.signal.aborted) {
-          return reject({ code: AppErrorCode.CANCELED });
-        }
+          if (abortController.signal.aborted) {
+            return reject({ code: AppErrorCode.CANCELED });
+          }
 
-        setTimeout(() => {
-          detectLogger.debug(`verifyAttendingDevice ${deviceId} attempt ${count + 1}`);
+          setVerifyingAttemptsCount(count);
           getMemberDevice(memberId, deviceId)
             .then((device) => {
               if (device?.attending) {
-                return resolve(device);
+                return device;
               }
-              return verifyAttendingDevice(memberId, deviceId, count + 1).then(resolve);
+              return verifyAttendingDevice(memberId, deviceId, count + 1);
             })
+            .then(resolve)
             .catch(reject);
         }, RETRY_BASE_DELAY_IN_MS * count); // increase the delay on each retry
+        setFetchDeviceTimeoutHandle(handle);
       });
     },
     [abortController],
@@ -98,27 +113,36 @@ const DetectDeviceBottomSheet = ({
 
   // fetch device info until it has been scanned by the probe
   const fetchDeviceInfo = useCallback(
-    async (count = 0) => {
+    async (count = 1) => {
       return new Promise((resolve, reject) => {
-        if (count > FETCH_DEVICE_INFO_MAX_ATTEMPTS_COUNT) {
-          throw new Error(t('devices.add.onFetchDeviceInfo.tooManyAttempts'));
-        }
+        const handle = setTimeout(() => {
+          pairLogger.debug(`fetchDeviceInfo attempt ${count}`);
+          if (count > FETCH_DEVICE_INFO_MAX_ATTEMPTS_COUNT) {
+            return reject(
+              new Error(
+                t('devices.add.onFetchDeviceInfo.tooManyAttempts', {
+                  count: FETCH_DEVICE_INFO_MAX_ATTEMPTS_COUNT,
+                }),
+              ),
+            );
+          }
 
-        if (abortController.signal.aborted) {
-          return reject({ code: AppErrorCode.CANCELED });
-        }
+          if (abortController.signal.aborted) {
+            return reject({ code: AppErrorCode.CANCELED });
+          }
 
-        setTimeout(() => {
-          detectLogger.debug(`fetchDeviceInfo attempt ${count + 1}`);
+          setFetchDeviveInfoAttemptsCount(count);
           getDeviceInfo()
             .then(({ device }) => {
               if (device) {
-                return resolve(device);
+                return device;
               }
-              return fetchDeviceInfo(count + 1).then(resolve);
+              return fetchDeviceInfo(count + 1);
             })
+            .then(resolve)
             .catch(reject);
         }, RETRY_BASE_DELAY_IN_MS * count); // increase the delay on each retry
+        setVerifyingTimeoutHandle(handle);
       });
     },
     [abortController],
@@ -135,9 +159,15 @@ const DetectDeviceBottomSheet = ({
       abortController.signal.addEventListener('abort', abortListener);
 
       (async () => {
+        setMacAddress(null);
+        setName(null);
+        setAddedDevice(null);
+        setVerifiedDevice(null);
+        setVerifyingAttemptsCount(0);
+        setFetchDeviveInfoAttemptsCount(0);
         const deviceInfo = (await fetchDeviceInfo().catch((error) => {
           if (error.code === 'ECONNABORTED' && error.message?.includes('timeout')) {
-            throw new Error(t('devices.add.onFetchDeviceInfo.unavailable', { cause: error }));
+            throw new Error(t('devices.add.onFetchDeviceInfo.unavailable'), { cause: error });
           }
           return Promise.reject(error);
         })) as ProbeDevice;
@@ -146,20 +176,32 @@ const DetectDeviceBottomSheet = ({
         const deviceName = deviceInfo.name || Device.deviceName || null;
         setName(deviceName);
 
-        const newDevice = await addMemberDevice(
-          authStore.user?.id as string,
-          {
-            name: deviceName,
-            macAddress: deviceMacAddress,
-            type: DeviceType.MOBILE,
-          } as ApiMemberDevice,
-        );
+        const [newDevice] = await Promise.all([
+          addMemberDevice(
+            authStore.user?.id as string,
+            {
+              name: deviceName,
+              macAddress: deviceMacAddress,
+              type: DeviceType.MOBILE,
+            } as ApiMemberDevice,
+          ),
+          // wait at least 1 second to let the user read the text
+          new Promise((r) => setTimeout(r, 1_000)),
+        ]);
         setAddedDevice(newDevice);
+        queryClient.invalidateQueries({
+          queryKey: ['members', authStore.user?.id, 'devices'],
+          exact: true,
+        });
 
         const attendingDevice = await verifyAttendingDevice(
           authStore.user?.id as string,
           newDevice._id as string,
         );
+        queryClient.invalidateQueries({
+          queryKey: ['members', authStore.user?.id, 'devices'],
+          exact: true,
+        });
         resolve(attendingDevice as never);
       })().catch(reject);
     });
@@ -173,15 +215,21 @@ const DetectDeviceBottomSheet = ({
       })
       .catch(handleSilentError)
       .catch((error) => {
-        noticeError(error, t('devices.add.onDetect.fail'));
+        noticeError(error, t('devices.add.onPair.fail'));
         resetAnimation();
       });
   }, [authStore]);
 
   const onCancel = useCallback(() => {
     abortController.abort();
+    if (fetchDeviceTimeoutHandle) {
+      clearTimeout(fetchDeviceTimeoutHandle);
+    }
+    if (verifyingTimeoutHandle) {
+      clearTimeout(verifyingTimeoutHandle);
+    }
     onClose?.();
-  }, [onClose, abortController]);
+  }, [onClose, abortController, fetchDeviceTimeoutHandle, verifyingTimeoutHandle]);
 
   const startAnimation = useCallback(() => {
     setAnimating(true);
@@ -203,6 +251,7 @@ const DetectDeviceBottomSheet = ({
   const resetAnimation = useCallback(() => {
     setLoopAnimation(false);
     setEndAnimation(false);
+    setAnimating(false);
     animation.current?.play(0, 0);
     animation.current?.reset();
   }, [animation]);
@@ -224,7 +273,7 @@ const DetectDeviceBottomSheet = ({
       bottomSheetRef.current?.close();
       Haptics.notificationAsync(Haptics.NotificationFeedbackType.Success);
       noticeStore.add({
-        message: t('devices.add.onDetect.success', {
+        message: t('devices.add.onPair.success', {
           name: verifiedDevice.name || verifiedDevice.macAddress,
         }),
         type: 'success',
@@ -244,10 +293,6 @@ const DetectDeviceBottomSheet = ({
             });
           }
         },
-      });
-      queryClient.invalidateQueries({
-        queryKey: ['members', authStore.user?.id, 'devices'],
-        exact: true,
       });
     }
   }, [verifiedDevice, noticeStore, t, bottomSheetRef, queryClient, settingsStore, toastStore]);
@@ -281,18 +326,31 @@ const DetectDeviceBottomSheet = ({
           <AppText
             exiting={FadeOutRight.duration(500)}
             style={tw`text-center text-xl font-bold tracking-tight text-slate-900 dark:text-gray-200`}>
-            {t('devices.add.detect.label')}
+            {t('devices.add.pair.label')}
           </AppText>
         )}
         {isAnimating && (
-          <>
+          <View style={tw`flex flex-col gap-1`}>
             {!macAddress && (
-              <AppText
-                entering={FadeInLeft.duration(1000)}
-                exiting={FadeOutRight.duration(500)}
-                style={tw`text-center text-xl font-bold tracking-tight text-slate-900 dark:text-gray-200`}>
-                {t('devices.add.onFetchDeviceInfo.pending')}
-              </AppText>
+              <>
+                <AppText
+                  entering={FadeInLeft.duration(1000)}
+                  exiting={FadeOutRight.duration(500)}
+                  style={tw`text-center text-xl font-bold tracking-tight text-slate-900 dark:text-gray-200`}>
+                  {t('devices.add.onFetchDeviceInfo.pending')}
+                </AppText>
+                {fetchDeviveInfoAttemptsCount > 1 && (
+                  <AppText
+                    entering={FadeIn.duration(1000)}
+                    exiting={FadeOutRight.duration(500)}
+                    style={tw`text-center text-xs font-normal text-slate-500 dark:text-slate-400`}>
+                    {t('devices.add.onFetchDeviceInfo.attempts', {
+                      count: fetchDeviveInfoAttemptsCount,
+                      max: FETCH_DEVICE_INFO_MAX_ATTEMPTS_COUNT,
+                    })}
+                  </AppText>
+                )}
+              </>
             )}
 
             {macAddress && !addedDevice && (
@@ -305,13 +363,25 @@ const DetectDeviceBottomSheet = ({
             )}
 
             {macAddress && addedDevice && (
-              <AppText
-                entering={FadeInLeft.duration(1000)}
-                style={tw`text-center text-xl font-bold tracking-tight text-slate-900 dark:text-gray-200`}>
-                {t('devices.add.onVerifyDevice.pending')}
-              </AppText>
+              <>
+                <AppText
+                  entering={FadeInLeft.duration(1000)}
+                  style={tw`text-center text-xl font-bold tracking-tight text-slate-900 dark:text-gray-200`}>
+                  {t('devices.add.onVerifyDevice.pending')}
+                </AppText>
+                {verifyingAttemptsCount > 1 && (
+                  <AppText
+                    entering={FadeIn.duration(1000)}
+                    style={tw`text-center text-xs font-normal text-slate-500 dark:text-slate-400`}>
+                    {t('devices.add.onVerifyDevice.attempts', {
+                      count: verifyingAttemptsCount,
+                      max: VERIFY_ATTENDING_DEVICE_MAX_ATTEMPTS_COUNT,
+                    })}
+                  </AppText>
+                )}
+              </>
             )}
-          </>
+          </View>
         )}
         <ReadMore
           numberOfLines={3}
@@ -333,7 +403,7 @@ const DetectDeviceBottomSheet = ({
             entering={FadeInLeft.duration(300)}
             exiting={FadeOutRight.duration(300)}
             style={tw`text-left text-base font-normal text-slate-500 dark:text-slate-400 w-full`}>
-            {t('devices.add.detect.description')}
+            {t('devices.add.pair.description')}
           </AppText>
         </ReadMore>
         <ReachableService pending={isReachingService} reachable={isServiceReachable} />
@@ -342,7 +412,7 @@ const DetectDeviceBottomSheet = ({
           style={tw`mt-2 w-full max-w-md self-center`}
           onPress={onStart}>
           <AppText style={tw`text-base text-black font-medium`}>
-            {t('devices.add.detect.start')}
+            {t('devices.add.pair.start')}
           </AppText>
         </AppRoundedButton>
         <Link
@@ -407,7 +477,7 @@ const ReachableService = ({
           <AppText
             entering={FadeIn.duration(300)}
             style={tw`text-left text-base font-normal text-slate-500 dark:text-slate-400`}>
-            {reachable ? t('devices.add.detect.reachable') : t('devices.add.detect.unreachable')}
+            {reachable ? t('devices.add.pair.reachable') : t('devices.add.pair.unreachable')}
           </AppText>
         </View>
       )}
@@ -415,4 +485,4 @@ const ReachableService = ({
   );
 };
 
-export default DetectDeviceBottomSheet;
+export default PairDeviceBottomSheet;